+++ antsor/lab_02/outline.h
@@ -16,7 +16,7 @@ public:
 	void set_vamount(int n);
 	
 	void move(int dx, int dy);
-	void scale(float kx, float ky, Point pm);
+	void scale(double kx, double ky, Point pm);
 	void rotate(int deg, Point pc);
 	
 	void draw(QGraphicsScene *scene);
